hookUpdate: moving some stuff around
